Created test method returnCoinTest.  Three quarters, one dime. Upon selecting CHIPS, Display should read "THANK YOU" and 35 cents should be returned as one quarter, one dime. Prices were converted to int for ease of calculating in small amounts this machine will have.

+++ src/com/wolfhaus/Coin.java
@@ -1,11 +1,13 @@
 package com.wolfhaus;
 
+import java.math.BigDecimal;
+
 public class Coin {
     protected int Size;
     protected int Weight;
-    protected double Value;
+    protected int Value;
 
-    public Coin(int coinSize, int coinWeight, double coinValue)
+    public Coin(int coinSize, int coinWeight, int coinValue)
     {
         this.Size = coinSize;
         this.Weight = coinWeight;
