+++ src/com/wolfhaus/Product.java
@@ -1,12 +1,15 @@
 package com.wolfhaus;
 
+import java.math.BigDecimal;
+import java.util.Currency;
+
 public class Product {
     protected String Button;
     protected String Name;
-    protected double Price;
+    protected int Price;
     protected int Count;
 
-    public Product(String name, double price, String button, int count)
+    public Product(String name, int price, String button, int count)
     {
         this.Name = name;
         this.Price = price;
